port: avoid breaking UCS10 FLL workaround
The UCS10 chip erratum affects many 5xx/6xx chips, and exhibits as an FLL
jump that generates a short-term clock error up to 12%.  See
http://www.ti.com/lit/pdf/SLAA489.  For chips with UCS, we leave the FLL
disabled except when explicitly recalibrating the DCO taps/modulation.  To
ensure that doesn't get broken, define a constant that should be used when
clearing the LPM bits on exit from interrupts and on task switches, and make
sure the value of that constant does not include SCG0 when UCS10 might
apply.

UCS7 is a similar bug that should also be eliminated through this practice.