Eliminate compile-time warnings
Unused variable and using char as an index cause gcc to complain.  Same with
providing a static declaration of a function in a header that's included in
files that don't provide a static definition of the function.  With these
fixed, the only complaints should be about casts between pointers and
integers, which might actually be relevant.  (The current ones are benign.)